@@ -17,8 +17,8 @@
     "prepare": "husky install",
     "updateDeipDeps": "node ./scripts/updateDeipDeps.js",
     "create:package": "npx yo ./packages/generators/generator-package",
-    "build:lib": "babel-node ./scripts/buildLib.js",
-    "watch:dev": "babel-node ./scripts/watchDev.js",
+    "build:lib": "babel-node scripts/buildProjectLib.js",
+    "watch:dev": "npm run build:lib && babel-node ./scripts/watchDev.js",
     "types:check": "tsc --noEmit"
   },
   "keywords": [],
@@ -74,7 +74,7 @@
     "sass-loader": "^10.2.1",
     "shx": "^0.3.4",
     "style-loader": "^2.0.0",
-    "typescript": "^4.5.4",
+    "typescript": "^4.7.2",
     "vue": "^2.6.14",
     "vue-jest": "^3.0.7",
     "vue-loader": "^15.9.7",
