@@ -4,7 +4,7 @@ import path from 'path';
 import fs from 'fs-extra';
 /* eslint-enable */
 
-const { rootPath } = require('./paths');
+import { rootPath } from './variables';
 
 const legacyExcludes = ['lib-crypto', 'RpcClient'];
 const lernaConf = fs.readJsonSync(path.join(rootPath, 'lerna.json'));
@@ -16,13 +16,26 @@ const getPackages = (forBuild = true) => lernaConf.packages
   .filter((pattern) => !pattern.includes('generators'))
   .filter((pattern) => !pattern.includes('portals'))
   .reduce((acc, pattern) => [...acc, ...glob.sync(pattern, { absolute: true })], [])
+
   .filter((p) => {
     const pArr = path.parse(p);
     if (forBuild) {
       return fs.existsSync(path.join(p, 'src')) && !legacyExcludes.includes(pArr.name);
     }
-
     return true;
+  })
+
+  .map((pkgPath) => {
+    const { name, dependencies } = fs.readJsonSync(`${pkgPath}/package.json`);
+
+    const platformDependencies = Object.keys(dependencies)
+      .filter((key) => key.includes('@deip') || key.includes('@casimir'));
+
+    return {
+      name,
+      path: pkgPath,
+      deps: platformDependencies
+    };
   });
 
 export {
